+++ apps/meteor/app/settings/client/lib/settings.ts
@@ -1,43 +1,39 @@
-import type { SettingValue } from '@rocket.chat/core-typings';
 import { Meteor } from 'meteor/meteor';
-import { ReactiveDict } from 'meteor/reactive-dict';
 
-import { PublicSettingsCachedCollection } from '../../../../client/lib/settings/PublicSettingsCachedCollection';
+import { watch } from '../../../../client/lib/cachedStores';
+import { PublicSettings } from '../../../../client/stores';
 import { SettingsBase } from '../../lib/settings';
 
 class Settings extends SettingsBase {
-	cachedCollection = PublicSettingsCachedCollection;
+	private readonly store = PublicSettings.use;
 
-	collection = PublicSettingsCachedCollection.collection;
-
-	dict = new ReactiveDict('settings');
-
-	get<TValue = any>(_id: string | RegExp, ...args: []): TValue {
+	override get<TValue = any>(_id: string | RegExp, ...args: []): TValue {
 		if (_id instanceof RegExp) {
 			throw new Error('RegExp Settings.get(RegExp)');
 		}
 		if (args.length > 0) {
 			throw new Error('settings.get(String, callback) only works on backend');
 		}
-		return this.dict.get(_id) as TValue;
-	}
 
-	private _storeSettingValue(record: { _id: string; value: SettingValue }, initialLoad: boolean): void {
-		Meteor.settings[record._id] = record.value;
-		this.dict.set(record._id, record.value);
-		this.load(record._id, record.value, initialLoad);
+		return watch(this.store, (state) => state.get(_id)?.value) as TValue;
 	}
 
 	init(): void {
 		let initialLoad = true;
-		this.collection.find().observe({
-			added: (record: { _id: string; value: SettingValue }) => this._storeSettingValue(record, initialLoad),
-			changed: (record: { _id: string; value: SettingValue }) => this._storeSettingValue(record, initialLoad),
-			removed: (record: { _id: string }) => {
-				delete Meteor.settings[record._id];
-				this.dict.set(record._id, null);
-				this.load(record._id, undefined, initialLoad);
-			},
+		this.store.subscribe((state) => {
+			const removedIds = new Set(Object.keys(Meteor.settings)).difference(new Set(state.records.keys()));
+
+			for (const _id of removedIds) {
+				delete Meteor.settings[_id];
+				this.load(_id, undefined, initialLoad);
+			}
+
+			for (const setting of state.records.values()) {
+				if (setting.value !== Meteor.settings[setting._id]) {
+					Meteor.settings[setting._id] = setting.value;
+					this.load(setting._id, setting.value, initialLoad);
+				}
+			}
 		});
 		initialLoad = false;
 	}