+++ apps/meteor/client/lib/constants.ts
@@ -1,3 +1,4 @@
 export const USER_STATUS_TEXT_MAX_LENGTH = 120;
 export const BIO_TEXT_MAX_LENGTH = 260;
 export const VIDEOCONF_STACK_MAX_USERS = 6;
+export const NAVIGATION_REGION_ID = 'navigation-region';