+++ apps/meteor/client/views/marketplace/AppDetailsPage/tabs/AppLogs/Filters/AppLogsFilter.tsx
@@ -3,6 +3,7 @@ import { useRouter, useSetModal } from '@rocket.chat/ui-contexts';
 import { Controller } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
 
+import AppsLogsFilterOptions from './AppLogsFilterOptions';
 import CompactFilterOptions from './CompactFilterOptions';
 import { EventFilterSelect } from './EventFilterSelect';
 import { InstanceFilterSelect } from './InstanceFilterSelect';
@@ -15,12 +16,13 @@ import { ExportLogsModal } from './ExportLogsModal';
 type AppsLogsFilterProps = {
 	appId: string;
 	expandAll: () => void;
+	collapseAll: () => void;
 	refetchLogs: () => void;
 	isLoading: boolean;
 	noResults?: boolean;
 };
 
-export const AppLogsFilter = ({ appId, expandAll, refetchLogs, isLoading, noResults = false }: AppsLogsFilterProps) => {
+export const AppLogsFilter = ({ appId, expandAll, collapseAll, refetchLogs, isLoading, noResults = false }: AppsLogsFilterProps) => {
 	const { t } = useTranslation();
 
 	const { control, getValues } = useAppLogsFilterFormContext();
@@ -96,11 +98,6 @@ export const AppLogsFilter = ({ appId, expandAll, refetchLogs, isLoading, noResu
 				</Box>
 			)}
 			{!compactMode && (
-				<Button alignSelf='flex-end' icon='arrow-expand' secondary mie={10} onClick={() => openAllLogs()}>
-					{t('Expand_all')}
-				</Button>
-			)}
-			{!compactMode && (
 				<IconButton
 					title={isLoading ? t('Loading') : t('Refresh_logs')}
 					alignSelf='flex-end'
@@ -129,8 +126,15 @@ export const AppLogsFilter = ({ appId, expandAll, refetchLogs, isLoading, noResu
 					{t('Filters')}
 				</Button>
 			)}
+			{!compactMode && <AppsLogsFilterOptions onExpandAll={openAllLogs} onCollapseAll={collapseAll} />}
 			{compactMode && (
-				<CompactFilterOptions isLoading={isLoading} onExportLogs={openExportModal} onExpandAll={openAllLogs} onRefreshLogs={refreshLogs} />
+				<CompactFilterOptions
+					isLoading={isLoading}
+					onExportLogs={openExportModal}
+					onExpandAll={openAllLogs}
+					onCollapseAll={collapseAll}
+					onRefreshLogs={refreshLogs}
+				/>
 			)}
 		</Box>
 	);