+++ packages/core-typings/src/Ajv.ts
@@ -2,6 +2,9 @@ import typia from 'typia';
 
 import type { ICustomSound } from './ICustomSound';
 import type { IInvite } from './IInvite';
+import type { IMessage } from './IMessage';
+import type { IOAuthApps } from './IOAuthApps';
+import type { IPermission } from './IPermission';
 import type { ISubscription } from './ISubscription';
 
-export const schemas = typia.json.schemas<[ISubscription | IInvite | ICustomSound], '3.0'>();
+export const schemas = typia.json.schemas<[ISubscription | IInvite | ICustomSound | IMessage | IOAuthApps | IPermission], '3.0'>();