+++ packages/core-typings/src/omnichannel/outbound.ts
@@ -8,11 +8,11 @@ export interface IOutboundProviderTemplate {
 	name: string;
 	language: string;
 	type: 'whatsapp' | 'email' | string;
-	category: 'AUTHENTICATION' | 'UTILITY' | 'MARKETING' | string;
+	category: 'authentication' | 'utility' | 'marketing' | string;
 	// Note: by default, the app will filter all the templates that are not APPROVED
-	status: 'APPROVED' | 'REJECTED' | 'PENDING' | string;
+	status: 'approved' | 'rejected' | 'pending' | string;
 	qualityScore: {
-		score: 'GREEN' | 'YELLOW' | 'RED' | 'UNKNOWN' | string;
+		score: 'green' | 'yellow' | 'red' | 'unknown' | string;
 		reasons: string[] | null;
 	};
 	components: Component[];
@@ -27,16 +27,16 @@ export interface IOutboundProviderTemplate {
 	partnerId: string;
 	externalId: string;
 	updatedExternal: string; // ISO 8601 timestamp
-	rejectedReason: string | null;
+	rejectedReason: string | undefined;
 }
 
 type Component = IHeaderComponent | IBodyComponent | IFooterComponent;
 
 // If we happen to have a different structure for this (since this could be a link or idk) we are gonna update this component type
 interface IHeaderComponent {
-	type: 'HEADER';
+	type: 'header';
 	// For UI: if the format is other than TEXT, it should include a media link
-	format?: 'TEXT' | 'IMAGE' | 'VIDEO' | 'DOCUMENT';
+	format?: 'text' | 'image' | 'video' | 'document';
 	text?: string;
 	example?: {
 		headerText?: string[];
@@ -44,7 +44,7 @@ interface IHeaderComponent {
 }
 
 interface IBodyComponent {
-	type: 'BODY';
+	type: 'body';
 	text: string;
 	example?: {
 		bodyText: string[][];
@@ -52,7 +52,7 @@ interface IBodyComponent {
 }
 
 interface IFooterComponent {
-	type: 'FOOTER';
+	type: 'footer';
 	text: string;
 }
 
@@ -73,12 +73,12 @@ export interface IOutboundMessage {
 	};
 }
 
-type TemplateComponent = {
+export type TemplateComponent = {
 	type: 'header' | 'body' | 'footer' | 'button';
 	parameters: TemplateParameter[];
 };
 
-type TemplateParameter =
+export type TemplateParameter =
 	| {
 			type: 'text';
 			text: string;