+++ ee/apps/presence-service/CHANGELOG.md
@@ -1,11 +1,105 @@
 # @rocket.chat/presence-service
 
-## 0.4.36
+## 0.4.37-rc.6
 
 ### Patch Changes
 
 - <details><summary>Updated dependencies []:</summary>
 
+  - @rocket.chat/core-typings@7.10.0-rc.6
+  - @rocket.chat/presence@0.2.37-rc.6
+  - @rocket.chat/core-services@0.10.0-rc.6
+  - @rocket.chat/model-typings@1.7.0-rc.6
+  - @rocket.chat/models@1.6.0-rc.6
+  - @rocket.chat/network-broker@0.2.16-rc.6
+  </details>
+
+## 0.4.37-rc.5
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+  - @rocket.chat/core-typings@7.10.0-rc.5
+  - @rocket.chat/presence@0.2.34-rc.5
+  - @rocket.chat/core-services@0.10.0-rc.5
+  - @rocket.chat/model-typings@1.7.0-rc.5
+  - @rocket.chat/models@1.6.0-rc.5
+  - @rocket.chat/network-broker@0.2.13-rc.5
+  </details>
+
+## 0.4.34-rc.4
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+  - @rocket.chat/core-typings@7.10.0-rc.4
+  - @rocket.chat/presence@0.2.34-rc.4
+  - @rocket.chat/core-services@0.10.0-rc.4
+  - @rocket.chat/model-typings@1.7.0-rc.4
+  - @rocket.chat/models@1.6.0-rc.4
+  - @rocket.chat/network-broker@0.2.13-rc.4
+  </details>
+
+## 0.4.34-rc.3
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+  - @rocket.chat/presence@0.2.34-rc.3
+  - @rocket.chat/core-services@0.10.0-rc.3
+  - @rocket.chat/core-typings@7.10.0-rc.3
+  - @rocket.chat/model-typings@1.7.0-rc.3
+  - @rocket.chat/models@1.6.0-rc.3
+  - @rocket.chat/network-broker@0.2.13-rc.3
+  </details>
+
+## 0.4.34-rc.2
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@7.10.0-rc.2
+  - @rocket.chat/presence@0.2.34-rc.2
+  - @rocket.chat/core-services@0.10.0-rc.2
+  - @rocket.chat/model-typings@1.7.0-rc.2
+  - @rocket.chat/models@1.6.0-rc.2
+  - @rocket.chat/network-broker@0.2.13-rc.2
+  </details>
+
+## 0.4.34-rc.1
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@7.10.0-rc.1
+  - @rocket.chat/presence@0.2.34-rc.1
+  - @rocket.chat/core-services@0.10.0-rc.1
+  - @rocket.chat/model-typings@1.7.0-rc.1
+  - @rocket.chat/models@1.6.0-rc.1
+  - @rocket.chat/network-broker@0.2.13-rc.1
+  </details>
+
+## 0.4.34-rc.0
+
+### Patch Changes
+
+- <details><summary>Updated dependencies [b25f05acd07762387fa45d67a1241b982c192f5d, fd32867fd4949bc2951a22075498ccb551cc6bbc, 17bca96ecbf23ea807aba2e6e8abc95ebd66b0d0, c7db598e9f3c2ad47f6a6be2a9ba7078533c245b]:</summary>
+
+  - @rocket.chat/model-typings@1.7.0-rc.0
+  - @rocket.chat/models@1.6.0-rc.0
+  - @rocket.chat/core-typings@7.10.0-rc.0
+  - @rocket.chat/core-services@0.10.0-rc.0
+  - @rocket.chat/presence@0.2.34-rc.0
+  - @rocket.chat/network-broker@0.2.13-rc.0
+  </details>
+
+## 0.4.36
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
   - @rocket.chat/core-typings@7.9.3
   - @rocket.chat/presence@0.2.36
   - @rocket.chat/core-services@0.9.15
@@ -19,7 +113,6 @@
 ### Patch Changes
 
 - <details><summary>Updated dependencies []:</summary>
-
   - @rocket.chat/core-typings@7.9.2
   - @rocket.chat/presence@0.2.35
   - @rocket.chat/core-services@0.9.14
@@ -33,13 +126,12 @@
 ### Patch Changes
 
 - <details><summary>Updated dependencies []:</summary>
-
-  - @rocket.chat/core-typings@7.9.1
-  - @rocket.chat/presence@0.2.34
-  - @rocket.chat/core-services@0.9.13
-  - @rocket.chat/model-typings@1.6.13
-  - @rocket.chat/models@1.5.13
-  - @rocket.chat/network-broker@0.2.13
+    - @rocket.chat/core-typings@7.9.1
+    - @rocket.chat/presence@0.2.34
+    - @rocket.chat/core-services@0.9.13
+    - @rocket.chat/model-typings@1.6.13
+    - @rocket.chat/models@1.5.13
+    - @rocket.chat/network-broker@0.2.13
   </details>
 
 ## 0.4.33