+++ ee/packages/network-broker/CHANGELOG.md
@@ -1,5 +1,65 @@
 # @rocket.chat/network-broker
 
+## 0.2.16-rc.6
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-services@0.10.0-rc.6
+  </details>
+
+## 0.2.16-rc.5
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+  - @rocket.chat/core-services@0.10.0-rc.5
+  </details>
+
+## 0.2.13-rc.4
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+  - @rocket.chat/core-services@0.10.0-rc.4
+  </details>
+
+## 0.2.13-rc.3
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+  - @rocket.chat/core-services@0.10.0-rc.3
+  </details>
+
+## 0.2.13-rc.2
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-services@0.10.0-rc.2
+  </details>
+
+## 0.2.13-rc.1
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-services@0.10.0-rc.1
+  </details>
+
+## 0.2.13-rc.0
+
+### Patch Changes
+
+- <details><summary>Updated dependencies [c7db598e9f3c2ad47f6a6be2a9ba7078533c245b]:</summary>
+
+  - @rocket.chat/core-services@0.10.0-rc.0
+  </details>
+
 ## 0.2.15
 
 ### Patch Changes
@@ -14,7 +74,6 @@
 ### Patch Changes
 
 - <details><summary>Updated dependencies []:</summary>
-
   - @rocket.chat/core-services@0.9.14
   </details>
 
@@ -23,7 +82,6 @@
 ### Patch Changes
 
 - <details><summary>Updated dependencies []:</summary>
-
   - @rocket.chat/core-services@0.9.13
   </details>