+++ ee/packages/presence/CHANGELOG.md
@@ -1,11 +1,48 @@
 # @rocket.chat/presence
 
-## 0.2.36
+## 0.2.37-rc.6
 
 ### Patch Changes
 
 - <details><summary>Updated dependencies []:</summary>
 
+  - @rocket.chat/core-typings@7.10.0-rc.6
+  - @rocket.chat/core-services@0.10.0-rc.6
+  - @rocket.chat/models@1.6.0-rc.6
+  </details>
+
+## 0.2.37-rc.5
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+  - @rocket.chat/core-typings@7.10.0-rc.5
+  - @rocket.chat/core-services@0.10.0-rc.5
+  - @rocket.chat/models@1.6.0-rc.5
+  </details>
+
+## 0.2.34-rc.4
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+  - @rocket.chat/core-typings@7.10.0-rc.4
+  - @rocket.chat/core-services@0.10.0-rc.4
+  - @rocket.chat/models@1.6.0-rc.4
+  </details>
+
+## 0.2.34-rc.3
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+  </details>
+
+## 0.2.36
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
   - @rocket.chat/core-typings@7.9.3
   - @rocket.chat/core-services@0.9.15
   - @rocket.chat/models@1.5.15
@@ -16,8 +53,7 @@
 ### Patch Changes
 
 - <details><summary>Updated dependencies []:</summary>
-
-  - @rocket.chat/core-typings@7.9.2
+- @rocket.chat/core-typings@7.9.2
   - @rocket.chat/core-services@0.9.14
   - @rocket.chat/models@1.5.14
   </details>
@@ -27,8 +63,7 @@
 ### Patch Changes
 
 - <details><summary>Updated dependencies []:</summary>
-
-  - @rocket.chat/core-typings@7.9.1
+- @rocket.chat/core-typings@7.9.1
   - @rocket.chat/core-services@0.9.13
   - @rocket.chat/models@1.5.13
   </details>